+++ src/features/reader/services/ReaderControls.ts
@@ -24,6 +24,7 @@ import { ReaderService } from '@/features/reader/services/ReaderService.ts';
 import type {
     PageInViewportType,
     ProgressBarPosition,
+    ReaderResumeMode,
     ReaderStatePages,
     ReadingDirection,
 } from '@/features/reader/Reader.types.ts';
@@ -192,9 +193,11 @@ class ReaderControlsClass {
         {
             doTransitionCheck = true,
             scrollIntoView = true,
+            resumeMode,
         }: {
             doTransitionCheck?: boolean;
             scrollIntoView?: boolean;
+            resumeMode?: ReaderResumeMode;
         } = {},
     ): void {
         const {
@@ -272,10 +275,12 @@ class ReaderControlsClass {
                 }
 
                 ReaderService.navigateToChapter(chapterToOpen, {
-                    resumeMode: getReaderOpenChapterResumeMode(
-                        isSpecificChapterMode || !keepRenderedChapters,
-                        isPreviousChapter,
-                    ),
+                    resumeMode:
+                        resumeMode ??
+                        getReaderOpenChapterResumeMode(
+                            isSpecificChapterMode || !keepRenderedChapters,
+                            isPreviousChapter,
+                        ),
                     updateInitialChapter: !keepRenderedChapters,
                 });
             } catch (error) {